edge_detection/requirements.txt
Pillow==9.4.0
tk==0.1.0
opencv-python==4.7.0.72